change to 5.0.3 release

+++ frameworks/core/common/platform_window.h
@@ -24,7 +24,8 @@
 #include "base/utils/noncopyable.h"
 
 namespace OHOS::Ace {
-using AceVsyncCallback = std::function<void(uint64_t, uint64_t)>;
+
+using AceVsyncCallback = std::function<void(uint64_t, uint32_t)>;
 class AceView;
 class RenderNode;
